feat: implement engagement directory bootstrap and auto-repair functionality for violin_guard
